Improved speed of Java Plotter
One of our users was having trouble with their local Python
installation for the Python based plotters. The Java plotter worked but
was very slow (a known problem). This commit includes a rewrite of the
file_xy code to eliminate the use of Vectors and substitute split for
the previous space-finding code. This increased the speed dramatically.