Checked length of object name before using as a key in bpy.data.objects
Somewhere in the building properties process, the object_name in the
active_mod_surf_regions is an empty string. This causes an error when
checking in "check_mod_surf_regions". It's not clear why this is
happening, but this commit implements a quick fix to simply check the
name first to see if it is empty. There's a considerable amount of debug
code inserted to try to trace the source of the problem, but it wasn't
obvious, so this quick fix was added. It might be wise to revisit this
issue at some time in the future.