Cleared run lists when changing Sim Runners ... and more on Data Model
The Queue runner and the previous sim runners differ in the lists they
use to keep track of processes. This can cause ugly display issues when
switching from one to another. This change implements a simple
"sim_runner_changed_callback" function that clears the both lists
whenever the simulation runner is changed. This change also displays
only the controls appropriate for the chosen simulation runner.