Only upgrade if mcell has been initialized, Make install shows CB ID
Saving a .blend file with uninitialized CellBlender properties was
mistaken for an RC3 file and an upgrade was attempted. Change to only
perform the upgrade if the mcell.initialized is true.

Have the make file show the cellblender_id.py file when installing as a
quick check of what was just installed.