Modified Java Plotter to allow for non-variable y scaling.
The previous version would scale the y range of the combined plot or of
each sub-plot based on the range of y values showing in the window (as
sampled on pixel values). This would cause the y range to shift as the
data might be shifted along the x axis. This is a useful mode, but it's
sometimes preferred that the y range remain fixed for the whole plot.
This version supports that with a menu-selectable option.