Have set_mol_glyph check for name, and organelle updates in test_suite
The set_mol_glyph was not checking if the molecule name was in the list
before setting the glyph. This shouldn't normally happen, but the check
is only when changing the glyph, so it's not a performance issue. This
was needed because the molecule name wasn't found to be set for some
cases in the newly added Organelle test case.

Note that the molecule color and glyph code needs more work to better
integrate the newer molecule definitions with the ability to read viz
data with no definitions (older method). Both should be retained so
CellBlender can use the new definitions while still being able to read
viz data sets where no molecules have been defined.