Consolidated surface regions into add_surface_region_to_model_by_normal
The add_surface_region_to_model_by_normal function compares the normal
of each face in the object with the requested normal and distance metric
to determine whether to add it to the region or not. This turns out to
be a convenient way to add surface regions to geometrical shapes like
cubes and icospheres. It was used to replace code in a number of tests.