MDL File Checking, modularization, and Sim Runner Tests
This version adds MDL file checking for a pass/fail test of each
simulation. The MDL for each test is saved as a single MDL file and is
then compared against a SHA1 of the expected MDL file (hard-coded into
the add-on). The results of the comparison are printed, and if they
don't match, then bpy.ops.wm.quit() is called to exit Blender.

Some more modularization was done to consolidate functionality.
Additionally, the tests were grouped into collapsable panels according
to the type of functionality being tested. An option was added to allow
just the checking of the MDL without actually running mcell. These
settings can be saved in the startup file so they'll be remembered after
pressing the reset button.

Simulation Runner tests were added to test each of the different kinds
of simulation runner. They all share the same simple model.