Better integration of Test Suite
It's not clear how the Test Suite should be integrated into CellBlender.

The previous version was completely stand-alone and needed to be
installed separately by copying the addon to Blender's addons directory.

This addon has not been included with other cellblender files in the
addons/cellblender directory because it is a separate addon. In fact,
this addon actually enables the CellBlender addon as part of its own
setting up for testing.

However, it is handy to have it installed via the make process, so a
separate line was added to the makefile to do so.

Finally, the previous version was named test_suite.py. This didn't
convey the fact that it was connected to cellblender (since it isn't
installed in the cellblender director), so it was renamed to
"cellblender_test_suite.py" to make the association clearer.