Moved rxn operators into cellblender_reactions.py and bug fix?
The while reactantStr[-1] in [quotes] for reactants and products was
failing when the reactant or product strings were of zero length. A
check was added prior to entering the while loop. Note that this error
probably didn't have much effect since reactions with empty reactants or
products would not be runnable anyway. But checking for the these cases
eliminated the exception being thrown.