Began adding cellblender_release.py, upgrade_data_model for bngl
Began work on refactoring the molecule release/placement code into a new
file named cellblender_release.py. It's not clear if this will also
contain the release pattern code (timing patterns).

Also found a bug in the commit from June 2nd, 2015 where mol_bngl_label
was added to the cellblender_molecules.py file. The buid methods were
properly implemented, but the upgrade_data_model function didn't contain
the code to add the 'mol_bngl_label' key/value into the data model.
Since the default for the new 'mol_bngl_label' was empty, the string ""
was chosen to initialize the value in the data model.

As a reminder, adding new properties involves four steps:

1. Create the property and initialize as normally required.

2. Provide a conversion of the property into a data model entry.

3. Provide a conversion of the data model entry back into the property.

4. Provide an "upgrade" to add/change the property from older data models.

That last step was missing, so when an older data model was opened, the
code that tried to perform step 3 found that there was no
"mol_bngl_label" key and failed. The data model upgrade is handled
separately from the building of the properties so that the properties
are never partially built by an incomplete data model. The upgrade
function will either fully upgrade the data model (which should make it
compatible with building the properties for the current version) or it
should fail which indicates that the existing properties should not be
changed or a broken .blend file may result.