Added proper data model upgrading for reaction output's new mdl_string.
A new property ("mdl_string") was added to the Reaction Output Group,
but it was not being added to older data models. This caused an error
when a .blend file with an older data model was opened.

There are 3 things to do with the data model when adding properties:

  1. Choose a version string to use and require it when building props.
  2. Put that version string into the data model when saving.
  3. Write the code to convert older data models to the new version.

All of these steps must be performed or your modifications WILL BE
incompatible with older .blend files. The most complicated step is the
updating of an older data model so it is compatible with the current
version. This is done in a series of steps:

  Is there any data model? If not, add one and call it V1.
  Is the data model V1? If so, upgrade it to V2.
  Is the data model V2? If so, upgrade it to V3.
  Is the data model V3? If so, upgrade it to V4.
   :
   :
  Is the data model V325? If so, upgrade it to V326.
  Is the data model V326? If so, upgrade it to V327.
  Is the data model V327? IF NOT SOMETHING WENT WRONG - QUIT NOW!!!

This "waterfall" logic isn't the only way to do this, but it's the
easiest. Any legitimate data model version can be handled and then
upgraded with this logic. The upgrades are a series of small patches to
get from one version to the next.

In the case of this commit, the "mdl_string" was expected to be in the
data model so it could be copied into the properties. Since older data
models didn't have that field, it needed to be added before it could be
used. So the "upgrade_data_model" function for that version should have
added an "mdl_string" to the data model with an appropriate default for
a blend file that had no "mdl_string" in the past (maybe None, maybe
""). Without that step, everything will work fine with your current
blend files, but it will crash when it looks for "mdl_string" in a
.blend file that doesn't have one.