Added a capsule primitive and a test case that uses it.
The capsule primitive is based on an icosphere that is split in half
along its "equator". The two halves are then separated by an amount
specified as the cylinder length (cyl_len). The gap is then filled with
either a single triangulated segment (may result in long thin triangles)
or a series of segments approximately the length of one of the icosphere
sides (each triangulated). The segmented version (subdivided_sides=True)
results in more triangles, but each triangle will have an area similar
to those in the icosphere caps.