Added wait loop to simulate waiting for MCell (or SimRunner) to complete
The "run_model" function in the test suite should be able to check on a
flag that's set by the sim runners (maybe?) to know when the simulation
is done running. Right now, each test just waits a fixed amount of time
which is clearly suboptimal based on the varying speeds of different
computers. It would be good to close the loop on the run time.