Added an "id" option for: "make id" to regenerate cellblender_id.py
It turns out that the CellBlender ID is regenerated whenever the make
file is invoked with any option since the the assignment of sources is
done by running cellblender_source_info.py which will regenerate that
file. However, this never had a dedicated make target to do this alone.
This change provides the target "id" which doesn't need to do anything,
but does print out the resulting CellBlender ID. This can be handy when
needing to verify the ID of a version without needing to build it or
install it or clean it.