Updated makefile in io_mesh_import_mcell_mdl
The makefile in io_mesh_import_mcell_mdl was installing to an older
version of Blender. This has probably not been a problem since that
makefile was not generally called with the "install" target. Instead,
our make system builds a zip file and the "install" target installs
from that zip file.