Changed references to utils to cellblender_utils.
The "utils.py" file was renamed to "cellblender_utils.py", but the
change wasn't made in all files. Since the install target of the
makefile didn't first delete the previously installed code, the old
utils.py file was still there and usable. This masked the problem when
installing via make install when an older installation was present.