Updated cellblender_id.py file. PLEASE READ.
In order to update the cellblender_id.py file, the makefile should be
run before making a commit. This is generally pretty easy to remember,
but there is a case where it's easy to forget!!!

When you're ready to push changes to GitHub, you may find that your push
is rejected because someone else has made changes since you last pulled.
The normal process at that point is to pull from the repository and
perform a "git merge" (while holding your breath hoping for no
conflicts). Whether you have to resolve conflicts or not, you should
always remember to re-run the makefile to ensure that the cellblender_id
file is updated to be consistent with the merge. If not, then the computed
cellblender_id will not match the cellblender_id in the file.

This particular commit was needed exactly for that reason. I had to
merge my previous work and had pushed without re-running the make file.