Added a dynamic geometry test case in a separate file: capsule_test.py
This dynamic geometry test code is fairly extensive - justifying its own
file. However, it is anticipated that it will be simpler when there is
more dynamic geometry support built into CellBlender. At that point it
should be merged into the main test suite file. For now, this test code
requires no changes to the CellBlender development branch. The Capsule
Maker addon should appear in the Scene Properties Panel at the bottom
(just like the older CellBlender addon).

The procedure for running is to enable both addons (CellBlender and
Capsule Maker). Uncheck the "Disable" flag in the Capsule Maker panel
so you can see the results of other setting changes in that panel. Make
any adjustments needed to get desired geometry over time, but leave it
in a good state to use for setting up a static CellBlender run (for
testing, all the original settings should be fine). Then recheck the
"Disable" flag so the addon doesn't continually generate dynamic geometry
while you're setting up the rest of your CellBlender run. Add the
"capsule" object as a CellBlender model object, but avoid changing any
of the names since they may still be hard-coded in this version.

Set up molecules, release sites, reactions, or anything else as needed.
You can run MCell simulations in this state. When you get your model set
up and running properly with static geometry, you should Decouple Export
and Run (Run Panel / Output/Control Options). That's because after
you've exported your MDL, the capsule addon will need to modify it for
dynamic geometry before you can run it. You should also manually set the
Start and End values in the time line to match your iterations. The
Capsule Maker addon uses those values to determine how much dynamic
geometry to generate. For a quick test, 100 iterations works well.

To run, you first save your blend file and export your model using the
Export CellBlender Project button. Then in the Capsule Maker panel, you
should press the "Set Path to MDL Files" button to let the addon know
where your MDL files are located. This should be in the standard place
relative to where you've saved your blend file. Then you press the
"Generate Dynamic MDL" button which generates the dynamic MDL files and
also modifies the Scene.main.mdl file to use the dynamic geometry. Note
that this ONLY works if you are using the Modular MCell MDL Files
option (currently the default). After Generating the dynamic MDL files,
you should be able to press "Run" (be sure that you've set your MCell
Binary to point to a dynamic geometry version of MCell!!!). After the
run, you should be able to Reload the Visualization Data and drag the
cursor in the time line to see the results. The capsule test simulates
a bacteria (ecoli?) growing, dividing, growing, dividing and so on ...
forever. The computation time per frame goes up with the number of cell
divisions so 2 or 3 divisions is a good range for testing. If you use
both volume and surface molecules, you should see them honoring the
boundaries of the dynamic geometry as it is moved through time.

As mentioned above, this addon was designed to work with the existing
CellBlender (development branch) without any special modification. As it
is, this should be able to be integrated into the test suite without too
much trouble. But it might also be a starting point for considering how
we want to implement dynamic geometry in CellBlender.