Moved a few operators to more appropriate locations.
The cellblender_operators.py file is now down to 250 lines, and the
cellblender_panels.py file is down to 75 lines. Much of the remaining
code deals with surface regions which will be the last to be done.

It's anticipated that the operators and panels files will be gone and
the remaining cellblender_properties.py file will become the new
cellblender_main.py file. The cellblender_legacy.py file will contain
code needed for backward compatibility as needed. This keeps the current
working code as clean as possible.