Renamed cellblender_properties.py to cellblender_main.py
The old cellblender_properties.py file now serves as the "main" file for
CellBlender. It defines the MCellPropertyGroup containing the different
components of the CellBlender application. It also contains the
CellBlenderMainPanelPropertyGroup which displays the user-selected
panels defined by the other components. This class also contains a few
of the somewhat global classes and functions used by other parts of the
CellBlender application.

For all of these reasons, the old cellblender_properties.py file was
renamed to be cellblender_main.py in this commit.

This version has passed all of the tests in the current test suite at
the MDL level. Not all tests were fully run in MCell.