Improved the capsule test for better integration with CellBlender
The capsule_test.py file tests dynamic geometry. The code generates a
dynamic mesh for Blender and can also generate a series of MDL files for
use by MCell. It can display it's own internally generated dynamic mesh,
but it can also read the MDL files that it generated and display them in
Blender as well (selected by a switch). It turns out that reading the
generated files gives a much faster display for dense geometries than
recalculating them on the fly.

This example is helping to flesh out the needs for supporting dynamic
geometry in CellBlender.