Use simulation time step for reaction step when left as blank.
The time step for reaction output (when left blank) had been zero. This
change checks the field for being blank and then uses the system time
step as the reaction output step when blank. This involved a change to
the io_mesh_mdl/export_mcell_mdl.py file for the logic, and it also
involved a change to the reaction output file to update the help text to
make this clear.