Cleaner handling of the export and run functions
Removed the decouple_export_and_run from the extra options in the
Preferences panel since it's now in the Run Simulation panel.

Added a poll method to the export operator to be consistent with
the Run operator. Exporting is disabled when it is locked out.
Note that exporting is also currently disabled when an MCell
binary is not found. This precludes generating MDL for other
purposes and should probably be allowed (although this can be
worked around by setting the MCell binary to any executable).

Updated the poll method for the run operator to disable the
run button when it is coupled with exporting and when exporting
is locked out.

Updated the drawing functions consistent with these changes.

Also fixed a bug in the handling of project_settings.base_name.
When a project was exported and then run separately (decoupled),
and saved in a blend file ... reopening and running again without
exporting caused an error. It appears that the project name was
still defaulted to "cellblender_project" in that case (even though
it should have been using the scene name. The "fix" in this commit
was to recompute the scene name (replacing space with underscore)
and then setting the project_settings.base_name to that name.