Updated Test Suite SHA1 Check Values for recent changes in 5bbf3b9c414
Commit 5bbf3b9c4147a558a2def5e79529c4ebec689bbe uses the system time
step value as the reaction time step when the reaction time step is not
specified (blank). This changes the output MDL for the following test
suite cases (listed by group in the current test suite):

  5. Surface Classes Test
  6. Capsule in Capsule Test

  1. Simple Molecule Count Test
  2. Release Time Patterns Test

  1. Lotka Volterra Torus - Diffusion Limited Reaction
  2. Lotka Volterra Torus - Physiologic Reaction
  3. Organelle Test
  4. E. coli MinD/MinE System

These tests previously generated MDL with "STEP=0" and now generate MDL
with STEP=[1e-06,1e-08,0.0002] (depending on the test case). Since the
test suite contained SHA1 values consistent with the "STEP=0" MDL, those
tests that now produce different values failed. The new MDL is correct,
so the SHA1 values in the test suite have been updated to match.