Generate output files named MDLString_#.dat to avoid name conflicts.
This is a temporary fix needed to generate multiple MDL String count
statements for the same molecule in different objects. The previous code
always generated the same file name for the same molecule even if it was
from different objects. This quick fix didn't construct a proper name
matching what was counted, but just added an incremented number to make
the file names unique.

The provision for including MDL in a CellBlender model is currently
under investigation for a more systematic solution.