Added Data Model to JSON converter functions.
Converting between a Data Model and JSON is a one line statement, but
these two functions are now provided to make it even easier.

Note, however, that the Data Model representation is native Python and
should contain all of the precision of native Python. JSON, however,
represents floating point numbers as decimal strings. From what I have
read, this is not a problem if the JSON implementation is done properly.
But for Python to Python communication, it's safest to stay with the
Pickled representation and only use the JSON conversion when interfacing
to non-Python code. If we envision communicating in a data-model-like
manner with MCell, it will probably be done through JSON, and these two
functions were provided for that purpose.