Added a Scripting Interface to support MDL and Python Scripting
The Scripting Interface consists of a new panel which allows the
specification of external and internal text files which are to be either
inserted in the MDL stream as MDL or executed as Python files to
generate MDL which is also inserted into the MDL stream. To make room
for the new panel (without increasing the number of buttons), the
Settings and Preferences panels have been combined into the new
"Settings & Preferences" panel. This Scripting Interface allows the
specification of scripts but the export code to actually insert the MDL
has not been written yet.

The Scripting Interface is a typical Blender list where Scripting Items
can be added or removed with the normal plus and minus buttons. Each
Scripting Item can specify an internal text file (inside the .blend file)
or an external text file (outside the .blend file). Internal files are
found by searching the list of files when refreshed and displaying that
list in a prop_search control. External files are found with the normal
file selection control. Each Scripting Item can also be specified as
either "MDL" or "Python". The exporter will copy the MDL files into the
MDL stream, and it will execute the Python files and direct their output
into the MDL stream. The insertion point for the resulting MDL can be
chosen as either before or after several of our main MDL sections. The
current options (right now) are:

  Before or After Parameters
  Before or After Initialization
  Before or After Molecules
  Before or After Reactions
  Before or After Geometry
  Before or After Output

These were taken from the MCell Quick Reference Guide (section 2.1)
which specifies the structure of an MDL file. We can add more sections
very easily as needed.

This capability should enable several new usage cases:

  Writing a 100% MDL simulation from within CellBlender
  Inserting custom MDL before or after any section
  Writing python to generate MDL for repetitive situations

Note that the output of all of these examples is just MDL and does not
create the new items inside CellBlender. While the custom Python could
generate new Blender properties when it is run, that's not a good use
scenario because the python will be called every time the simulation is
run. In order to have user python generating new CellBlender properties,
we will probably want to use a different mechanism to invoke it.

Note also that the implementation of the custom Python is expected to
get a copy of the Data Model when it is invoked. This will allow it to
easily investigate what's in the current simulation and generate MDL
consistent with the current model. For example, if the custom Python
code were to assign a specific surface class to objects of the name
"Circle.xxx", then it could do so based on examining the data model and
finding out how many "Circle.xxx" objects are in the model.

Finally, it's also possible that the custom Python could be used to
modify a temporary data model. This would allow it to completely change,
delete, or regenerate any section of the model. The changes to the Data
Model could be applied only to the exporting of the MDL and not change
the properties inside Blender itself. In other words, the Python code
would essentially "patch" the Data Model before it is used to generate
MDL, but those patches would only apply to that run and not change the
Data Model stored with the .blend file.

The ability to add custom MDL and insert custom Python opens up many new
opportunities, and it may take some time before they're all apparent.