Fixed all_warnings (had used all_notifications), Removed script writes
It apppears the the output of individual warnings had been keying off of
the "all_notifications" flag when writing the MDL output.

Also, removed the scripting.write_scripting_output calls. These
generated comments in the MDL code where custom MDL scripts would be
placed. However, since the scripting update was moved to a separate
branch, these comments served no purpose and also made the resulting MDL
files incompatible with the test suite. So to preserve the functioning
of the test suite, these comments to the MDL have been commented out.

Ideally, the scripting output will not change the MDL if there are no
scripting files chosen for inclusion. So the existing test suite SHA1
values for their MDL files will not have to be recalculated when (or if)
scripting is included.