Added safety setting required to enable Python Scripting.
This change is intended to support today's decision to allow Python
scripting when explicitly enabled by the user.

It is desired that this setting be explicitly disabled whenever a new
.blend file is opened - from whatever source. For that reason, a new
load post handler has been added to the cellblender_simulation.py file.
That new load post handler (named "disable_python") is added to the
load_post handler list by __init__.py. So whenever a new .blend file is
opened, it should be called. It, in turn, sets "enable_python_scripting"
to false in the "mcell.run_simulation" property group. This is the flag
that will be checked before running any python code when a CellBlender
project is exported to be run by MCell.