Ran the make file to update the CellBlender ID in cellblender_id.py.
The hash code in cellblender_id.py is a "fingerprint" of the code that's
being run, and it can be re-confirmed at run time by recomputing it on
all of the CellBlender files. So if a user has a problem, we can
confirm that their source code hasn't been modified and we can also
track down the GIT commit for that version by searching the history of
the cellblender_id.py file:

  git log -S'a2d73b58115437e1b642e4937e' -- cellblender_id.py

That will show the actual commit that generated that id ... IF the make
file was run to update that file prior to the commit.