Data Model to MDL functionality
This commit includes a preliminary version of "data_model_to_mdl.py".

Ideally, this would be able to take a data model and export a complete
set of MCell MDL files. Ideally, this would be a stand-alone function
without the need for any blender-specific data. In practice that might
prove to be too limiting.

This commit also includes a relatively trivial change to allow exporting
of MDL from the data_model_tree.py program. Right now the output MDL is
put in the file "data_model.mdl" - which will be overwritten every time
it's run.