Exported partitions from Data Model to MDL; Missing z_start bug fixed.
The following changes arose from work on exporting MDL directly from a
data model rather than from properties.

Updated some of the string representations saved in the data model to
match the export format used when exporting MDL. This mostly involved
explicitly using the "%g" format rather than relying on "str()".

Fixed a bug in the storage of partitions to the data model. The
"z_start" value was being saved as "x_start" in the data model. An
upgrade to the data model was implemented to handle this for older
.blend files.

Exported partitions in data_model_to_mdl.py exporter.

Note that the Data Model to MDL exporter exports parameters as
expressions. This has been an option in CellBlender, and that can
account for differences in the exported MDL. When making strict
comparisons between normal export and the Data Model to MDL exporter,
it's important to enable exporting parameters as expressions.