Generate Parameter Data Model in dependency order for MDL export
General Parameters in the Data Model are represented by a list of
dictionaries (one for each parameter). The list of parameter
dictionaries in the data model does not have to be in dependency
order. However, in order to generate valid MDL the order must be
correct.

data_model_to_mdl.py *could* contain the logic required to examine
each parameter for dependencies and export in the correct order, but
this would either require considerable duplication of code or the
inclusion of the parameter code in data_model_to_mdl. Since it's
handy to have data_model_to_mdl.py work even outside of CellBlender,
and since the parameter system within CellBlender already contains
the logic to sort the parameters into dependency order, this
implementation uses that code to store the parameters in the data
model list in dependency order. If CellBlender generates the data
model list in correct order, then data_model_to_mdl.py doesn't need
to do any resorting when it generates the MDL. The primary down side
to this approach is that exporting to a data model and then importing
it back into CellBlender might change the order of the parameters in
CellBlender's GUI list. Since the order in that list has never been
defined to have any meaning, this is a reasonable approach for now.