Check for existing colors before adding - temporary fix
The current color system should be reworked. It would be good to store
molecule color in the data model so that exchanges of data models would
share the same original colors. However, Blender also has its own color
system which has been used to assign molecule colors. These two have not
yet been completely reconciled. One of the side effects recently noticed
is a proliferation of duplicate colors when data models are read into
and out of properties. To curb this (without actually reconciling how
we're handling colors) this commit simply checks whether a color exists
or not before adding it to the data model or to the properties. This may
not be what we eventually want to do, but it does cut down on the rapid
growth of colors in a .blend file that's been updated many times.