Added a preliminary gnuplot plotter (not included in makefile yet).
This is a preliminary data plotter designed to use gnuplot. This version
always plots cos(2*x) and is only being used to see if gnuplot can be
used as a CellBlender data plotter. For that reason, the makefile has
not been updated to include this plotter in the CellBlender zip file.