Plots all files in the same window ... one after the other.
This version runs gnuplot with a pipe to stdin, stdout, stderr. It then
issues a series of plotting commands (one for each file) to stdin, and
gnuplot plots them one at a time (replacing the previous plot each time).

While this isn't consistent with what the other plotters do, it does
make a nice time-varying display.