Finished a working version of GnuPlot and improved Simple Plot legend
The GnuPlot plotter turned out to be very fast for showing many plots on
a single page (Fick's Law, for example). Also upgraded Simple Plot to
honor the no-legend option to avoid filling the entire plot with legend
information.

Note that while the GnuPlot plotter is in the GIT repository, it's not
currently included in the makefile or the resulting zip file until it's
approved for general use.