Allow Defaults in Molecule Data Model
The original idea for the data model was an internal representation that
fully captured the Blender properties for the entire MCell model. In
that case, it was reasonable to expect that a data model would contain a
full set of all properties.

However, since we are now exploring the possibility of using the data
model as a specification interface, it makes sense to allow defaults for
values. This changes implements that for molecule fields by only
changing a molecule field if its value exists in the data model. In the
past, such a partially defined data model would create errors.