Add separation between object list and object properties.
The object list allows the addition and removal of Blender objects to
the CellBlender model. It does not actually select objects, and it does
not reflect the object that is selected.

The object properties (shown below the list box) reflect the properties
of the object selected in the 3D window and not the object selected in
the model objects list (above).

At some point, we would like selection in either the list or the 3D view
to change the other (selecting in 3D selects in list, and selecting in
the list selects in the 3D view). We have not been able to make that
work within Blender, so a separator has been added to better show that
the properties do not reflect the selected object in the list.