Creation of a synthetic "ALL" region for assigning surface classes
Issue #94 reports the need to handle "ALL" in the context of molecules.
The "ALL" problem has also cropped up in the "Assign Surface Classes"
portion of the code.

The quick solution for the "Assign Surface Classes" case was to add a
boolean flag that specified "All Surfaces". When that flag was checked,
the region selection box was simply not drawn, and the "ALL" was
exported in the MDL.

While the extra check box works, it's not quite as streamlined (for the
GUI) as adding an "ALL" region to the region selector list.

Unfortunately, the region selector list is drawn automatically by
Blender from a real properties list. So to get an "ALL" option means
really adding an "ALL" region to the list of regions in each object.
That's what this code does.

Of course, that creates other problems (creating the synthetic "ALL"
region at startup, keeping the synthetic "ALL" region from being
exported to MCell or to the CellBlender data model, not showing the
synthetic "ALL" region in the region assignment panel, and so on).

This branch is an attempt to address these problems in support of an
"ALL" region for assigning surface classes.

Note that this branch is not completed yet. In particular, none of the
data model upgrading has been implemented, so .blend files opened with
this branch may not upgrade properly, and .blend files saved from this
branch may not be properly upgradable to future versions.