Updated test suite to use add_surface_region_to_model_by_normal for all
The add_surface_region_to_model_by_normal test suite function accepts a
vector and a minimum dot product that's used to select faces to be
included in the region. That function has a default vector of (0,0,0)
which is used to signify that all faces be included in the region.

However, that differs from using the ALL MDL keyword. The test suite had
defined a function named "add_surface_region_to_model_all_faces" which
simply called the other function with a normal vector of all zeros. The
result was an MDL file that specified all the faces of the object, but
did not use the "ALL" keyword.

This commit changes the existing tests so that they use the "normal"
function so that the "all" function can really use the "ALL" keyword.

This commit does not change any results of the test suite.