Modified the test suite to actually test the "ALL" region keyword.
The previous surface classes test defined "all" regions by simply
enumerating all of the regions producing MDL that looked like this:

   DEFINE_SURFACE_REGIONS
   {
     a_reg
     {
       ELEMENT_LIST = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]
     }
   }

The region was then referenced as: ai[a_reg]

This commit continues to define most surface regions as described above,
but uses the "ALL" keyword for the "ai" object rather than defining a
region for it. So the previous "DEFINE_SURFACE_REGIONS" MDL block is NOT
produced and the "region" is referenced simply as: ai[ALL].

This was verified by checking the resulting MDL with diff.